+++ api/src/Claim.ts
@@ -1,9 +1,9 @@
+import { HexString } from '@polkadot/util/types';
 import { ISubmittableResult } from '@polkadot/types/types';
 import { SubmittableExtrinsic } from '@polkadot/api/types';
 
 import { ClaimValueError } from './errors';
 import { GearTransaction } from './Transaction';
-import { Hex } from './types';
 
 /**
  * Claim value from mailbox
@@ -13,7 +13,7 @@ export class GearClaimValue extends GearTransaction {
    * Submit `claimValueFromMailbox` extrinsic
    * @param messageId MessageId with value to be claimed
    */
-  submit(messageId: Hex): SubmittableExtrinsic<'promise', ISubmittableResult> {
+  submit(messageId: HexString): SubmittableExtrinsic<'promise', ISubmittableResult> {
     try {
       this.extrinsic = this._api.tx.gear.claimValue(messageId);
       return this.extrinsic;