+++ api/src/Code.ts
@@ -1,10 +1,14 @@
-import { Bytes, Option } from '@polkadot/types';
+import { Bytes, Option, Vec, u8 } from '@polkadot/types';
+import { HexString } from '@polkadot/util/types';
 import { ISubmittableResult } from '@polkadot/types/types';
 import { SubmittableExtrinsic } from '@polkadot/api/types';
+import { u8aToHex } from '@polkadot/util';
 
-import { CodeMetadata, CodeStorage, Hex } from './types';
-import { generateCodeHash, validateCodeId } from './utils';
+import { CodeMetadata, CodeStorage } from './types';
+import { generateCodeHash, getIdsFromKeys, validateCodeId } from './utils';
+import { CodeDoesNotExistError } from './errors';
 import { GearTransaction } from './Transaction';
+import { getGrReply } from './wasm';
 
 export class GearCode extends GearTransaction {
   /**
@@ -14,7 +18,7 @@ export class GearCode extends GearTransaction {
    */
   async upload(
     code: Buffer | Uint8Array,
-  ): Promise<{ codeHash: Hex; submitted: SubmittableExtrinsic<'promise', ISubmittableResult> }> {
+  ): Promise<{ codeHash: HexString; submitted: SubmittableExtrinsic<'promise', ISubmittableResult> }> {
     const codeHash = generateCodeHash(code);
     await validateCodeId(codeHash, this._api);
 
@@ -36,7 +40,7 @@ export class GearCode extends GearTransaction {
    * ### Get code storage
    * @param codeId
    */
-  async storage(codeId: Hex): Promise<CodeStorage> {
+  async storage(codeId: HexString): Promise<CodeStorage> {
     return this._api.query.gearProgram.codeStorage(codeId) as unknown as CodeStorage;
   }
 
@@ -44,7 +48,7 @@ export class GearCode extends GearTransaction {
    * ### Get static pages of code
    * @param codeId
    */
-  async staticPages(codeId: Hex): Promise<number | null> {
+  async staticPages(codeId: HexString): Promise<number | null> {
     const storage = await this.storage(codeId);
     return storage.isSome ? storage.unwrap().staticPages.toNumber() : null;
   }
@@ -53,11 +57,28 @@ export class GearCode extends GearTransaction {
    * ### Get all ids of codes uploaded on connected chain
    * @returns array of code ids uploaded on chain
    */
-  async all(): Promise<Hex[]> {
-    const keyPrefix = this._api.query.gearProgram.metadataStorage.keyPrefix();
-    const codeMetadata = await this._api.rpc.state.getKeys(keyPrefix);
-    const codeIds = codeMetadata.map((key) => '0x' + key.toHex().slice(keyPrefix.length)) as Hex[];
+  async all(): Promise<HexString[]> {
+    const prefix = this._api.query.gearProgram.metadataStorage.keyPrefix();
+    const keys = await this._api.rpc.state.getKeys(prefix);
+    const codeIds = getIdsFromKeys(keys, prefix);
 
     return codeIds;
+  }
+
+  async metaHash(codeId: HexString): Promise<HexString> {
+    const code = (await this._api.query.gearProgram.originalCodeStorage(codeId)) as Option<Vec<u8>>;
+    if (code.isNone) {
+      throw new CodeDoesNotExistError(codeId);
+    }
+
+    const metahash = await getGrReply(code.unwrap().toHex(), 'metahash');
+
+    return u8aToHex(metahash);
+  }
+
+  async metaHashFromWasm(wasm: Buffer | ArrayBuffer | HexString | Uint8Array) {
+    const metahash = await getGrReply(wasm, 'metahash');
+
+    return u8aToHex(metahash);
   }
 }