+++ api/src/create-type/decodeHex.ts
@@ -1,7 +1,8 @@
-import { Hex } from '../types';
+import { HexString } from '@polkadot/util/types';
+
 import { TypeInfoRegistry } from './TypeInfoReg';
 
-export function decodeHexTypes(hexTypes: Hex | Uint8Array) {
+export function decodeHexTypes(hexTypes: HexString | Uint8Array) {
   const typeInfoReg = new TypeInfoRegistry(hexTypes);
   return typeInfoReg.getTypes();
 }