+++ api/src/types/interfaces/message/extrinsic.ts
@@ -1,15 +1,16 @@
 import { H256 } from '@polkadot/types/interfaces';
+import { HexString } from '@polkadot/util/types';
 
-import { GasLimit, Hex, Value } from '../../common';
+import { GasLimit, Value } from '../../common';
 import { PayloadType } from '../../payload';
 
 export interface IMessageSendOptions {
-  destination: Hex | H256;
+  destination: HexString | H256;
   payload: PayloadType;
   gasLimit: GasLimit;
   value?: Value;
 }
 
 export interface IMessageSendReplyOptions extends Omit<IMessageSendOptions, 'destination'> {
-  replyToId: Hex;
+  replyToId: HexString;
 }