+++ api/src/utils/validate.ts
@@ -1,7 +1,8 @@
 import { BN, u8aToBigInt } from '@polkadot/util';
 import { u128, u64 } from '@polkadot/types';
+import { HexString } from '@polkadot/util/types';
 
-import { GasLimit, Hex, Value } from '../types';
+import { GasLimit, Value } from '../types';
 import { GearApi } from '../GearApi';
 import { ValidationError } from '../errors';
 
@@ -14,8 +15,8 @@ export function validateValue(value: Value | undefined, api: GearApi) {
     value instanceof Uint8Array
       ? u8aToBigInt(value)
       : value instanceof u128 || value instanceof BN
-        ? BigInt(value.toString())
-        : BigInt(value);
+      ? BigInt(value.toString())
+      : BigInt(value);
 
   if (bigintValue > 0 && bigintValue < existentialDeposit.toBigInt()) {
     throw new ValidationError(`Value less than minimal. Minimal value: ${existentialDeposit.toHuman()}`);
@@ -28,14 +29,14 @@ export function validateGasLimit(gas: GasLimit, api: GearApi) {
     gas instanceof Uint8Array
       ? u8aToBigInt(gas)
       : gas instanceof u64 || gas instanceof BN
-        ? BigInt(gas.toString())
-        : BigInt(gas);
+      ? BigInt(gas.toString())
+      : BigInt(gas);
   if (bigintGas > api.blockGasLimit.toBigInt()) {
     throw new ValidationError(`GasLimit too high. Maximum gasLimit value is ${api.blockGasLimit.toHuman()}`);
   }
 }
 
-export async function validateCodeId(codeId: Hex, api: GearApi) {
+export async function validateCodeId(codeId: HexString, api: GearApi) {
   if (await api.code.exists(codeId)) {
     throw new ValidationError('Code already exists');
   }